+++ tests/e2e/helpers/rules.js
@@ -10,7 +10,6 @@
  * @param {string} [options.js] Inline JS (may not execute under Playwright CSP)
  * @param {string} [options.css] Inline CSS
  * @param {string} [options.html] Inline HTML
- * @param {Array<{path: string, type?: string, ext: string}>} [options.files]
  */
 export function makeRule({
   selector = '127\\.0\\.0\\.1',
@@ -20,29 +19,12 @@ export function makeRule({
   js = '',
   css = '',
   html = '',
-  files = [],
 } = {}) {
   return {
     selector,
     enabled,
     onLoad,
     topFrameOnly,
-    code: { js, css, html, files },
-  };
-}
-
-export function remoteFile(origin, name, ext) {
-  return {
-    path: `${origin}/${name}`,
-    type: 'remote',
-    ext,
-  };
-}
-
-export function localFile(filePath, ext) {
-  return {
-    path: filePath,
-    type: 'local',
-    ext,
+    code: { js, css, html },
   };
 }